+++ front/admin/prestop.sh
@@ -1,1 +1,2 @@
-curl -X POST "http://localhost:3000/$PRESTOP_SECRET/prestop"
+#!/bin/sh
+curl -X POST "http://localhost:3000/api/$PRESTOP_SECRET/prestop"
