+++ front/global.d.ts
@@ -1,0 +1,6 @@
+declare global {
+  // eslint-disable-next-line no-var
+  var wakeLocks: Set<string> | undefined;
+}
+
+export {};
