+++ front/tsconfig.json
@@ -17,9 +17,11 @@
     "downlevelIteration": true,
     "paths": {
       "@app/*": ["./*"]
-    }
+    },
+    "allowUnreachableCode": true
   },
   "include": [
+    "global.d.ts",
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
